Extends KeyboardHandler behavior for non-tty stdin.
