+++ delphyne_demos/demos/keyop.py
@@ -10,67 +10,17 @@ A keyop demo.
 # Imports
 ##############################################################################
 
-from select import select
-
-import atexit
-import sys
-import termios
-
 import delphyne.trees
 import delphyne.behaviours
 
 from delphyne_gui.utilities import launch_interactive_simulation
 
 from . import helpers
+from . import keyboard_handler
 
 ##############################################################################
 # Supporting Classes & Methods
 ##############################################################################
-
-
-class KeyboardHandler(object):
-    """A keyboard-interrupt poller. Allows users to read a keyboard
-    input with a non-locking behavior making use of the select function,
-    available on most *nix systems.
-
-    This class is based on the work done by Frank Deng, available on GitHub
-    as part of a set of python tools released under the MIT licence:
-    https://github.com/frank-deng/experimental-works/blob/master/kbhit.py .
-    """
-
-    def __init__(self, input_stream=None):
-        if input_stream:
-            self.input_stream = input_stream
-        else:
-            self.input_stream = sys.stdin
-        # Save current terminal settings
-        self.file_descriptor = self.input_stream.fileno()
-        self.new_terminal = termios.tcgetattr(self.file_descriptor)
-        self.old_terminal = termios.tcgetattr(self.file_descriptor)
-        # New terminal setting unbuffered
-        self.new_terminal[3] = (self.new_terminal[3] &
-                                ~termios.ICANON & ~termios.ECHO)
-        termios.tcsetattr(self.file_descriptor,
-                          termios.TCSAFLUSH, self.new_terminal)
-        # Support normal-terminal reset at exit
-        atexit.register(self.set_normal_term)
-
-    def set_normal_term(self):
-        """Resets to default terminal settings."""
-        termios.tcsetattr(self.file_descriptor,
-                          termios.TCSAFLUSH, self.old_terminal)
-
-    def get_character(self):
-        """Reads a character from the keyboard."""
-        char = self.input_stream.read(1)
-        if char == '\x00' or ord(char) >= 0xA1:
-            return char + self.input_stream.read(1)
-        return char
-
-    def key_hit(self):
-        """Returns True if a keyboard key has been pressed, False otherwise."""
-        key_hit, _, _ = select([self.input_stream], [], [], 0)
-        return key_hit != []
 
 
 def demo_callback(behaviour_tree, keyboard_handler, time_step_seconds):
@@ -142,7 +92,7 @@ def main():
         time_step=sim_runner_time_step
     )
 
-    keyboard = KeyboardHandler()
+    keyboard = keyboard_handler.get_keyboard_handler()
 
     # We add it as a step callback because the runner
     # gets stuck in a while loop until it's unpaused.
